# api/swagger/swagger.yaml
definitions:
  github_com_copolio_namegpt_internal_namegpt_entity.DomainName:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      queryID:
        type: integer
      updatedAt:
        type: string
    type: object
  github_com_copolio_namegpt_internal_namegpt_entity.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_copolio_namegpt_pkg_dto_request.CreateUser:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  github_com_copolio_namegpt_pkg_dto_request.RecommendDomainNames:
    properties:
      description:
        type: string
      previousResults:
        items:
          type: string
        type: array
      userId:
        type: string
    required:
    - description
    type: object
  github_com_copolio_namegpt_pkg_dto_request.SimilarDomainNames:
    properties:
      keyword:
        type: string
      user:
        type: string
    required:
    - keyword
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v0/ping:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Pong!
      tags:
      - v0
  /api/v0/users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create user request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_copolio_namegpt_pkg_dto_request.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/github_com_copolio_namegpt_internal_namegpt_entity.User'
      summary: Creates a user.
      tags:
      - v0
  /api/v0/users/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/github_com_copolio_namegpt_internal_namegpt_entity.User'
      summary: Gets a user info by name.
      tags:
      - v0
  /api/v1/domains/recommendations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Generate domain name recommendation
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_copolio_namegpt_pkg_dto_request.RecommendDomainNames'
      produces:
      - application/json
      responses:
        "201":
          description: domain
          schema:
            items:
              type: string
            type: array
      summary: Generates domain names in server sent event.
      tags:
      - v1
  /api/v1/domains/similar-names:
    post:
      parameters:
      - description: Generate similar domain name
        in: body
        name: q
        required: true
        schema:
          $ref: '#/definitions/github_com_copolio_namegpt_pkg_dto_request.SimilarDomainNames'
      produces:
      - application/json
      responses:
        "200":
          description: domain names
          schema:
            items:
              $ref: '#/definitions/github_com_copolio_namegpt_internal_namegpt_entity.DomainName'
            type: array
      summary: Get similar domain names given input
      tags:
      - v1
swagger: "2.0"